Prevent HotelRulesFragment crash
If you rush to the hotels legal information page from the hotel booking overview, while
the "calculating taxes and fees" is occuring, we don't yet have the cancellation policy,
but try to display it in HotelRulesFragment. Check that the cancellation policy exists
before attempting to display it, and just leave the TextView blank in that case.

We could prevent the opening HotelRulesActivity altogether until this info is available, but
this seems like a simple enough fix for a kind of edge-casey thing.