1083: Removed window bg from tablet search
It was causing massive amounts of overdraw.

We also are keeping the map visible at all times.  This is because
before it was useless what we were doing anyways (hide() would turn
off on rotation), and it makes the animation still look good even
without the window background (otherwise you get blurry random pixels)