Refactor HotelGalleryActivity
It had 3 inner class definitions, some of which were tightly coupled to eachother. I
put some of them in their own classes and decoupled them. This just sort of follows
the pattern of other similar components in the codebase, so it should be easier to
read now. The main impetus was that diagnosing a bug was sort of tricky just because of
how strange the code looked.