1358 NPE on HotelGalleryActivity (tablet)
The problem is that the ViewPager adapter was depending on the
Gallery to get the number of pics, but the Gallery does not get
assigned until after the ViewPager adapter starts to undergo some
of it's overrides. This decouples the gallery from the ViewPager
adapter and just refers to the Property's media to retrieve the count.