1369 hotels calendar rotate bug
Whew, this was a doozy to track down. Basically, now that we allow null Dates for
the checkIn and checkOut times, we need to make sure we keep track of these null
values when serializing/deserializing via the JSONable interface. Otherwise, when
we construct the HotelSearchParams after a rotation, it gets a default stay set
which include both a checkIn and checkOut date, leading to a whacky state. Preserve
these null dates because it means something now.

This is also a PSA to use JSONUtils when fromJson/toJson-ing classes that implement
the JSONable interface.