1368 reflect updated total via coupon in hotel breakdown
Susanna pointed out that I was showing the discount from the coupon in green, but also
updating the per night rate (which seems to be where the discount is taken from). Now,
show the per night rate as the original rate, and the total price + coupon discount is
where the coupon differences will be.