String.equals()/equalsIgnoreCase() heresy
Ok, so this fixes a bug in getPosLocale() where the wrong locale is selected and we
display links, text in the wrong language. You can test this by switching your POS
to expedia.com.hk (leaving the device language as English) and then clicking the
Privacy Policy link from the info screen. You will notice that the webpage that loads
is in Chinese because we fail to determine the correct PosLocale. For some reason
"en".equalsIgnoreCase("en") was returning false. So its not String literals, but those
are the values of the Strings, and their contents appear to be the same in the debugger.

Switching to use compareIgnoreCase() properly identifies the Strings as the same, and we
can go on our merry way. I have now idea why equalsIgnoreCase()/equals() return false. The
only thing that comes to mind is one String is from Locale.getDefault().getLanguage() and
the other String is parsed from a JSON file in the assets folder.