Simplified ItinListView.FooterView
Removed the slightly hacky FooterVisibilityView, which was being used to
make the ListView not ignore footer view sizes. A different way to have the
ListView not ignore the footer height is to add a footer view that can
change size (from 0 to whatever) when we need it. But I've found that you
can't just change the size of the view, it must be a layout that WRAP_CONTENT
and have its content change size. I guess this is still hacky but at least
we are not creating a new FooterView every time we expand a row.