Fix overlapping text bug on itin cards
The summary and detail TextViews sometimes yield overlapping TextViews
and cause the text to be unreadable. Scroll around and click the itins
with the itin7@expedia.com account to test. This ensures that reused
container has its children removed before adding a new view to it. The
thing that confuses me, is that I would not think the container would have
children inside of it due to ItinCardDataAdapter.getItemViewType and also
ItinCard.bind() reuse logic...but it appears to happen.

Auditors: dlew