Reorganize hotel/flight traveler/payment classes
Basically, it is very confusing to look at these files because there are
four Activity-s, and all of their fragments, for the traveler and payment
entry screens for both flights and hotels and the files are all so very
similar. This commit attempts to organize the code better, so that it is
easier to read (also, follows https://team.mobiata.com/wiki/Android_Style_Guide)

Things done in this commit:
-Follow the style guide
-Organize methods together in a more logical way
-Make public methods private
-Remove unnecessary code (lifecycle overrides that just call super, newInstance methods creating args but not using them, etc...)

No actual code has changed, but this commit sets the stage for reducing what seems to be
a lot of duplicate, copy/pasted code so that this code is easier to maintain, change, read.