1401: Show full gridlayout rows
Now we set the left columns to stretch and the right column has a set
size.  This allows the right column to align right, while letting the
left column wrap intelligently (if there is too much text).

This is a quick hack solution before we have time to do a full-featured
solution, as set size columns are not a good idea.