Sync with manager using temporary vars
The problem was that the sync was taking so long that it was possible for the
adapter and the listview to get out of sync.  This caused the app to crash.

Now we calculate the new adapter data in separate variables and update
everything at once, so there's no chance that the adapter and the listview will
desync.