1449: Fixed shading of icon ImageView
After some investigation it turned out that we did seem to be setting
the alpha correctly, but that it was not being preserved for some
reason during the animations.  I've created an AlphaImageView which
does the alpha in its own onDraw(), which seems to preserve the alpha
of the View much better than the View's alpha setting itself.

This also fixes a side problem of the ImageView popping in/out of
full/shade alpha during the animation.

(This probably has something to do with NineOldAndroids' ViewHelper,
 but I don't want to delve so deep into the issue that I have to fix
 NOA).