1458: Fixed ItinListView crash
By overriding getAdapter() I was sometimes messing with how the ListView
worked.  It would think it would be getting the ListAdapter (that wraps
the header/footer Views) but instead it would only get our Adapter, which
would have a different count for the # of items.  This would cause
ListView to be very sad.