AccountManager part2 - We now use Authenticator for all of our sign in needs.
As of this commit, our login extenders are broken. I was also haveing some inflating xml binary blah blah blah issues.
The good news is that our login state in the app should match our login state in the AccountManager, and the app
now uses the same methods as the Add Account button in settings which is pretty cool