AccountManager part4 - Fix parcelable bug.
So AccountManager cant handle custom type parcelables, because it pulls them out
of the bundle and it cant unmarshal classes it doesnt know about. This is a known bug.
Thus we moved to this sort of bastardized Factory style of serializing and unserializing
loginextenders. https://code.google.com/p/android/issues/detail?id=55252#c1