1428: Use new HotelTextSections for description sections
The code works the same as before, the data just comes from a different
source.

We use a backwards-compatible method of getting the HotelTextSections
that can still fall back to parsing the hotel description HTML.  This
is because only trunk currently has the hotel text attributes; once
those are in place we should remove HotelDescription and related classes.