Move content provider authority declarations to XML
In effort to support multiple build flavors, we need to modify the
content provider authorities such that content providers do not conflict
and prevent installation of more than 1 flavor of our project.

The docs are relatively unclear on the topic of defining the authority
as a String resource vs. raw string. In practice, this works, so I am
moving forward with it.