Updated WebViewFragment to include a dialog mode.
This is a little wonky because typically fragments are intended to be
dialogs OR inline, but not both. We work around that shortcoming, and ensure
things work ok. We also remove the requirement for attaching to a WebViewListener activity.
Future work might involve including a loading indicator in the dialog, and better styling