1506: Added disk caching of flight route data
We don't want to have to requery this every time we boot the app.
When we load the data, we check to see if we have saved data.  If
we do, we load it.  If it's not stale, we use it instead of doing
a new download.

We are also careful to clear out the existing data if anything ever
changes (like POS or locale).