Relaunch app if locale changes
There's just a bunch of things we don't want to have to deal with, for example
we might start hitting a different langid due to the locale change midway
through.  So let's just go back to the launcher.

Also got rid of logic that is now invalid in the phone search activity.