1473 user reviews API migration pt 3
Now that we are using the new URS API, get rid of a bunch of custom logic that
we are no longer following. This is mostly related to the custom set of languages
and locales, of reviews, that we would request based on POS. The URS API simply
accepts a locale, and then returns the reviews that we want.