LoginFragment - support already logged in.
Previously we could never get to the log in screen while logged in.
Account manager changes that. We now show a simple account button if we
are already logged in, giving the user the option of logging out, and
logging in with a different user. The only gotcha here, is that calling
User.signOut invalidates the FB session, so we need our listener to ignore that.