Fixed build issue with themes
The root of the problem was 1506.  I needed to define a new
dropDownListViewStyle but it was already defined to change the non-spinner
popups.  As a result we had two competing themes, depending on dynamic POS
data.  I've changed the code such that the Fragment's theme changes based
on which airport selector we're displaying.