Fix VSC flavor crash wrt GCM
GCM requires some permissions defined as raw string in the AndroidManifest that are dependent on the
app's packageName. In order to properly support GCM in different flavors, we would need to move out
the GCM <permission> and <receiver> declarations into flavor-specific AndroidManifest files, and not
the default AndroidManifest. This means we would be forced to build EBad with gradle, which we are not
quite ready for yet.

In light of this, and the fact that GCM code doesn't run on a VSC build, we just prevent interacting with
GCM on the VSC build.