Fix versioning for alpha builds
Now that we define versionCode in build.gradle, gradle does not refer to
the AndroidManifest versionCode at all. Moreover, versionCode is configurable
only for productFlavors and not buildTypes. HockeyApp used to require that
versionCode increments on every build, but recently started supporting a
meta-data value. This commit effectively switches over to using the HockeyApp
buildNumber meta-data tag to upload new HockeyApp builds/