1577 - Speed up User.signOut.
The bottleneck was with the ItineraryManager.clear and Notification.deleteAll calls.
Because these are not directly related to logged in state, we move them to a background
thread in the new User.signOutAsync method. Before a user with a bunch of itins would
take 1-2 seconds (really bad for UI) to logout. Now we are a more reasonalbe 50 - 300ms.
I am pretty open to critisizm for this change, so please phab-fire-away.