1687: When GWallet promo rejected, still allow purchase
The problem was one of timing; the last time we'd update view
visibilities was when the coupon error was being displayed.  However,
this would take place while the download key was still valid, meaning
that the system would think the coupon was still being downloaded.

Now we make sure to update view visibilities after the user closes
the error message from the GWallet promo.