Further protect Google Wallet enabled status
Now it's a private variable in WalletFragment that cannot be
changed by any of its subclasses accidentally.  This makes it even
more impossible to re-enable Google Wallet by accident.

Eventually we should switch it to be a positive boolean, aka
mGoogleWalletEnabled, as that makes more sense; but for this point
release I don't want to risk mucking up the logic.