1327 When a user logs out from itins, dont be shitty.
In the past users would see a logout dialog, click logout, the gui would sort of hang for a minute,
and then they would be logged out. The problem was that during the logout operation, the user could
still click outside of the dialog (acting as a dialog cancel) and then click on an Itin all before the
UI refreshed at all because the UI thread was being bogged down by User.signOut.

Solution, we move to using the User.signOutAsync, so the logout operation is relatively quick,
We tell the adapter to disregard what is in ItinManager for a minute, and just pretend to be empty
and then do a sync after everything is all the way logged out. Probably too complicated for this
type of bug, but it solves our problem and enhances the User.signOutAsync call a bit by allowing
for a callback.