1538 Fix missing email checkout crasher.
This ended up being slightly more than a bug fix. In times past we depended on the programmer
to set the email address in BillingInfo correctly (but were still worried about it being wrong).
Because we know the various places email addresses can come from, we now always look through those
places in order of importance to ensure that we not only have an email address, but a valid one that
came from the right place (the logged in user for example). If no valid email addresses are present
we now notify the user via toast, and update view states. This is not great, but is much better
than crashing just as the user was about to give us their money.