Remove some buggy code from PhoneSearchActivity
broadcastSearchCompleted was getting invoked at the end of onCreate, which was touching
the HotelListFragment before it was ready. As it turns out, this code is two years old
and doesn't look to be required anymore (especially when it causes a crash).
broadcastSearchCompleted will get invoked if it needs to in onResume, otherwise the
Fragment can handle drawing itself because all of the pertinent data is in Db anyway.