BillingInfo email refactor.
In the past, we were relying on the programmer to call Db.getBillingInfo().setEmail() to set
the email to be used at checkout time. This was fine when there were 2 modes (logged in, and logged out)
but now we have added google wallet, and stuff got weird. So, for the time being we are using
BillingInfo.setEmail() to represent the MANUALLY ENTERED email address. When the user gets to the slide
to checkout phase of checkout, we copy the appropriate email address (from User,GWallet,BillingInfo) to
the billingInfo object we are going to use at checkout.

Future work will involve removing most calls to BillingInfo.getEmail() with a more universal method that
takes User and GWallet state into account. We will also want to never set the billingInfo email address
from anywhere other than manual entry.