1807. VSC Disable app widgets.
Since the user has to open the app at least once before the widgets are disable we need to make sure that the right VSC assets are shown.
The new attr referencing via the theme was not working for widgets, hence had to rename the file and have it be replaced in res_vsc.