1664 Carte Si handling.
As it turns out we have no idea how to detect Carte Si. One day however, we might have the
technology. That said, we now display generic card icons for brands that we dont have specific
assets for. When we do figure out Carte Si (and other card brands) the generics will be displayed