Only use Bitmaps with HeaderBitmapDrawable
We were trying to load the placeholder resource into the
HeaderBitmapDrawable, under the assumption it was a Bitmap.  That is
definitely not always the case; sometimes it's an XML drawable, which
would cause our app to blow up.

Now we only use HeaderBitmapDrawable if we know, for sure, that we're
dealing with a Bitmap.