Remove unused disk cache methods from PhoneSearchActivity
There were plenty of remnants of code that was being used to save and reload cached
search parameters and search results. While the code was being invoked to save the
data, it was never being loaded. Remove this code in preparation of actually using
the saved data in a future commit.