1856: Let HeaderBitmapDrawable use a placeholder Drawable
The problem was that, by skipping HeaderBitmapDrawable, we skipped a
bunch of the things we wanted to do (like rounded corners and gradients).
We need to be able to accept a generic Drawable (as UrlBitmapDrawable
only deals with Bitmaps), so now it supports one as well.

We try to upgrade to the best "source" as quickly as we can, be it a
placeholder or the actual Bitmap we ant to use.

Also, ItinCard images are again entirely driven by HeaderBitmapDrawable.
We don't want to switch away from it or some rows will look funny.