Use asset-based TZ Provider for Joda
It seems there is an issue with using ClassLoader.getResourceAsStream().
It causes any JAR to balloon the memory usage when you try to do it.
Joda time uses it by default to load all of its TZ data, which is stored
within the JAR itself.

Luckily you can provide your own TZ provider, so I've added one which uses
the asset system to retrieve TZ data instead.  It is almost exactly
the same as how Joda does it now, except that it uses assets instead of
getResourceAsStream().

We could rip the TZ data out of our local Joda JAR but it's not taking up
enough space to be worth the extra effort right now; it will just
remain unused.