Added new SearchParams
This will be a SearchParams that can handle all types of searches; it makes
more sense in the long run as we start doing package searches.

Note that it's not intended to always be *able* to do all searches; it might
only have enough information for a hotel search, but not a flight search,
for example.