Changed launch activity to adjustPan
This is to avoid two problems:

- Animation translation: if the activity starts resized (with the keyboard
  already displayed), then we have no way of measuring the size of the
  fragments when they are at normal size.

- Performance: When the app resizes, this causes a huge jank.  Removing this
  resize smooths out performance even when the keyboard is involved.

It has one drawback: it means you can't actually scroll to the bottom of the
suggestions list most of the time.  I consider this alright; the bottom of
the list will probably not contain the one you're looking for.  If this is
a big problem ultimately we can figure out something to do (like add a footer
to the ListView which lets you scroll down), but I could not figure out a way
to avoid the problems resizing creates otherwise.