Fix hockeyapp id
There was a disconnect between the hockeyId from jenkins/hockeyIdentifier
used for uploading to the hockeyApp servers and the hockeyId from Codes.java
used for in-app updating, crash reporting. We were uploading to Expedia
Beta but checking for updates from Expedia Alpha. that leads to weirdness.

This will fix auto-updating in-app for the Expedia Beta/3.4