1552: Fixed crash caused by errant lower-casing
It turns out that in some locales (Turkish for example), lower
casing the letter 'I' results in a weird, dotless i.  As a result,
we could not find that country identifier when we tried to reflect
it.

Since all of this is internal anyways, we now explicitly use the
English locale for lower casing all strings.