Began work on backstack management, submodes (e.g. controller has own state), and added the hotels filter fragment.
This is where we begin to test the robustness of the childFragmentManager system. Not only do we have the Controllers maintaining their own
internal state on top of the global state, but I added the first ValueAnimator for mode changing.
This is of coarse reversable, which is baller. Futhermore, the global state, and the substate are
maintained on rotation. Probably some other stuff too.

Having the controllers have their own states makes things a little harry, because we end up having
n global states * m controller states. This will probably be the focus of some thought and refactoring
in the near future