1917 - Remove auto syncing, instead just make stuff sync on the checkout screens.
Basically we only need synced data for checkout stuff, and we were already doing one
sync when we made it to the checkout screen. We now just make sure to sync again if
our duration (30 mins) has passed and the user resumes the screen. Again it is so very
unlikely that onResume would not be called if a user was using that screen that this
is good enough.