Tablet Results - First crack at add hotel animation.
This is a first attempt. At the very least it will be refactored to better handle things like
rotation and back press handling. Futhermore we are currently keeping the moving hotel that is
being added above the trip pane, which I imagine could cause headaches down the road. Also could
use some performance tweaking. I think we are starting to hit the limit for things we should have
using hardware layers. Still, it looks nice, and I think the way we pass between the hotels controller
and trip controller is pretty solid.