1940: Don't re-search on resume by accident
We had an issue where a simulated error would cause the app to want to do a
new search on resume.  However, this is a problem if we did a succesful search,
because the state would never reset; as a result, we'd start a new search
erroneously when you resumed the search activity.  Now we properly reset the
state.