Tablet Results - Refactor the way we add hotel trips.
We remove a global state, and instead let the HotelsController be in charge of adding
its hotel to the trip. This makes a lot more sense in terms of where things need to be
and when. We should also see a performance boost on animations because now our add trip
animation is actually being controlled by the old hotel list, which is badass. This is
marking a pretty good version of how we will move trips from their respective controllers
to the trips bucket