Tablet Results - Flights animations v1.
This allows the flights things to fly around, and gives a slightly better feel for the flights flow
but the code is not up to snuff. There are some rotation bugs, and more importantly I dont think
the right things are being controlled by the right fragments. I wanted to push to get this off of
my local box. Expect future iterations.