Tablet Results - rework flight details fly around animations.
The old solution was pretty cracked out. The new solution lets the details fragment do the work, but is configured by
the controller fragment. Futher it has a much more sane order of operations, and generally just works better. We have
not yet implemented all of the transitions, nor have we set up hardware layers and other goodies, but this is now on the
right track