Tablet Results - add flights to trip animation v1.
This will still need some polish. There is atleast one rotation bug I am aware of. We also have an addToTrip
fragment for flights, and no similar fragment for hotels. It would be nice to maintain paridy, so that needs
some thought. On the plus side we added ScreenPositionUtils, so now doing handoff style animations shoudl be
considerably easier, as we can easily pass coordinates from one view to another. This is another v1 commit.
So it goes.