Better error handling for reviews screen failures.
This code was present before the refactor but was entirely removed
during the refactor because it was not possible to open the reviews
screen and have a tab not show reviews (we wouldn't let the user
open the reviews activity). Now that we filter out reviews on a tab
by tab basis, it is possible for a tab to have no reviews present.
I lifted most of this code from the dead.