Itin sharing. Replace the sharableURL to /m instead of /api
The api returns the sharableUrl in the form of /api/trips/shared. But this is the link that is to be shared to any users.
We will instead replace it with /m/trips/shared, so that if the user clicks on the link without the app installed it will get them to the mobileWeb's shared itin landing page.