Tablet filter widget consolidation effort
The flights and hotels tablet filters use the same looking UI component. Doug and I are trying
to unify this common code. We've created "CheckBoxFilterWidget" to do just this. It might not
be the perfect class, but it does cut down on unnecessary code duplication and remains pretty
simple despite supporting both hotels and flights.

Some discussion about it here:
https://phabricator.mobiata.com/D37