Introducing SwipeOutLayout a viewgroup for swiping out content.
This viewgroup must have 2 subviews, one for content, and one for indicating the swipeout.

The swipe direction gets defined by a custom attribute, allowing us to use this view for swiping north,south,east,west.

This viewgroup should be useful for several tablet results screens, notably swiping south to remove an item from
hotel comparison, and swiping west to remove an item from your trip.

This viewgroup defines the ISwipeOutListener which will report swipe starts,aborts,and completions along with
the percentage of the current swipe.