1579 - use nearby airports suggestions API
This is present on the phone (and v1 tablet) flights search UI. In it's current
form, there should be no behaviorial change in the data that is returned. Ultimately,
we are hoping the suggest API returns a really good list that we can use. Right now
we just use sort by DISTANCE and filter client side on classification.