Tablet Results - Refactor mega commit. Removed ITabletResultsController in favor of new interfaces.
The functionality from ITabletResultsController has been replaced by the more granular interfaces:
 IStateProvider,IStateListener,IMeasurementListener,IMeasurementProvider,IBackManagable

These are implemented in such a way that they could be used in other cases. They more easily allow for the use of Helpers such as:
BackManager,MeasurementHelper,StateListenerHelper,StateListenerLogger

Basically ITabletResultsController was a mad house which was not easy to follow or set up. The new
interfaces intend to make things more obvious, and easier to implement.

This is a paridy commit, so the idea here was to just get things to the way they were before. More
helpers, and more organized and crunched code are still to come. This may not solve all of our problems
but I think it is a step in the right direction towards more understandable code.