Added airports to SearchParams
There are cases where the origin/destination airport may be a different
location from the hotel search location.  For example, if you are doing
a current location search, the nearest airport may be a hundred miles
away but you don't want to search for hotels there, you want them here.