Move CalendarTouchUtils to tests
It was living in TestUtils, but, compiling a library project
for a test project is not supported and would require special
hacks to get working. CalendarTouchUtils is the only Java code
that is being used from TestUtils, so we just port it over to the
main tests src directory and call it a day.