2128: Filter out all minor airports from results
I tried to think of a time when we'd want to suggest an airport where
you can't fly to it but couldn't, so we're doing the easy thing and
getting rid of them at parse time.  If we ever do need minor airports
then we could remove it later.

(cherry picked from commit e12d79ca1765777cd037cab675ef60fb739a21d4)