2092. ItinSharing. Reverting minicard changes.
The design came up a new way to display the imported sharedItin in mini-card view.
The info line in the mini-card remains the same, just as a normal itin.
Going forward the icon would no longer be a silhouette, but will be an image with traveler initials painted on top of the image.
For now i will leave the silhouette icons in there.