Added our own version of getRelativeTimeSpanString()
The DateUtils version is inherently buggy because it does not calculate
the number of days well at all.  I've tried hacking together solutions
but none of them seem to work right because it keeps using the default
timezone (since you can only use milliseconds).

The easiest solution at this point seems to be to implement it ourselves,
using Joda Time instead of milliseconds.  This means we have to copy
a bunch of strings from the Android source but the implementation is
correct now.