Automatically update the widget every X hours
The number of hours depends on whether the app is in a high energy
state at the moment.

Whenever switching to a higher energy state, we check if the last
update time was previous to the high energy update schedule; that
way we avoid the problem where switching the device on/off causes
the app to keep re-scheduling without ever doing a new search.