2160: Removed errant argument change
Here's what happened: we branched master from 3.4.1.  Then
master changed the number of args in two strings.  That was then
sent for the next loc dump, which was put on master, then cherry-
picked into 3.4.1 to fill out a few missing strings.

Unfortunately having the wrong # of arguments causes the app to blow
up.  So this reverts those particular string changes.