2135 rewards points show when first loading flight checkout
Joel mentioned that the architecture recently changed when he refactored the
animation on this screen. As such, the FlightCheckoutFragment exists before
the rewards points are returned in the create trip call from the price fragment.
Now we notify the activity the create trip call has been completed, and we can
rebind the checkout fragment to show the points.