2178: Initiate searches when date changes
In high power state, this is achieved via ACTION_TIME_TICK; we check
every minute if the date has changed.

In a low power state, none of this is done; but if you switch to
high power state, then we verify if the date is still correct, and if
not then we initiate a new search right there.