2181 - Fix itin sharing bug for sharing multiple package items.
Packages have multiple items, so a user may choose to first share their flight, and then decide they want to share
their hotel info. Great. The problem here is that packages have a top level uniqueId, so in the past if we shared
two package trips, the newer of them would simply squash the older, and then we would lose thing we first shared.
This is now fixed.