2201 - We should now be showing shortened urls for all trip types.
In this rather large commit, we standardized the way we get trip sharing urls via the ItinShareInfo class.

Basically anyplace we can have a share url, we can also have a shortened share details url. This stuff is all over the place,
because we can share a flight leg, a flight leg in a package, a different flight leg in a package, a hotel, a hotel in a package,
blah blah blah, basically we cant associate our share info with top level objects, and that makes this all crazy, but this way
we are at least a little bit standardized.