Show laoding indicator when importing shared itin to empty itin list
We needed to create a dummy Trip to conform with existing behavior, similar to
how guest trips work.

The loading indicator does not come up if there are itins already in app because
we've done work to make the progress indicator less aggressive in the past.

Noah said he would create a story for this but I never saw one so I'm just quietly
slipping this loading indicator in...