Use Trip.getItineraryKey() where necessary
Trip.getItineraryKey() is the definitive call to get a unique key on a Trip, mostly for use
within the ItinManager. We recently started using this method to account for shared itins
where the unique piece of data is the shareable url. This commit cleans up the Trip "key"
usages.

Auditors: brennan