2213 - Fix FT push notification header strings.
We are looking through the args array for the destination, and when we added all the FT strings we never updated
the logic for finding which of the arguments was the destination ( it was arg 0 for almost all EB push strings )
but now we actually look at the push key and figure out which argument represents the destination.

Currently FT is doing a lookup based on the fhid returned with the push message, but because we have the information
we need in the arguments, this way makes some sense for now.