2217: Fixed action bar items not showing up
With ABS, invalidating the options menu instantly updates the header.  We were
previously depending on the ability to enqueue an action bar update, then call
a sync, and have the synced items be available by the time the header updated.

Now we just switch the order so the sync happens before we try to update the
action bar.