2258 fix 2.x visual bug for VIP access badge
I suspected it was something to do with setting alpha on a view that was invisible, so I set
the view to gone instead. I'm not entirely sure what is happening but, I suspect it is related
to the Canvas stuff that goes on in AlphaImageView, and it only happens on 2.x. An added bonus is
setting this view to GONE will remove it from the hierarchy when it is not needed.