Reverting commit 964c0c0378998abf903de5fdcd1bdd62809f182f.
The commit being reverted was put in place to merge shared itin packages. E.g. we would receive a shared url for
a flight, and then later we would receive a different shared url for a hotel in the same package, because the trip
contained both items, we would have to merge them into a single trip. We now key on share url for shared things so
no merging is required and this commit becomes useless.