2244 - Fix expanded itin text position on rotation bug.
Lucky for us we were already addressing this problem for 2.x phones - basically if we had an itin that shows the icon even when expanded
and we rotated, the itin title text would sit underneath the icon and look bad. Now if we are in expanded mode, and we rotate, we double
check if the title and header text will be showing, and ensure that the title text is translated as to not sit underneath our icon.