2269 - Fix login rotation bug.
This ended up being more of a state problem - if we were already logged in, but we were still loading itins
and we rotated, we would check if we were logged in, and then we would set the state to LOGGED_IN which just
displays a login button, which isnt what we wanted, but we never noticed it because the edittexts were getting
restored after this, which fired the TextWatchers which made the thing think we were in the regular log in state
so we displayed the login button to the user. When they clicked that thing, it resulted in a crash (as we were
already logged in, and already loading our itins).

To fix the problem we dont add our TextWatchers until onViewStateRestored (after the edittexts have restored their
values) and we make sure not to set our state to LOGGED_IN if we are still loading itins after a login