2150 - Fix login state issue.
The issue: Imagine we do a network request in the bg (itin manager), while that network request is happening the user hits logout,
this removes the users login cookies from our cookie store. Eventually the network request from earlier finishes, and we save the
cookies that are returned. These returned cookies still have our login info in them, and squash our cookie store that has the correct
logged out state. BAD BAD BAD.

The solution: Before we get our cookie store to do a network request, we check if we are logged in. We do the request. When it returns
we check and see if the login state has changed. If the login state has changed, we just redo the network request with a new set of
cookies that better reflect the current login state. If our new cookies dont have permission, so be it, we just fail in the normal way.