Facebook ItinSharing for cars and activities.
Without the link the FB post doesn't show up, yes - FB Sdk sucks and we know it.
So we will be linking the desktop mobile app page for anything but hotels and flights.
Also added the default image links for FB post to use when sharing cars and everything else.