Flight filter bug fix
There was some weirdness going on with the Filters because of how I would generate
the set of airports for use within the FlightFilter. Originally, I would just look
at the departure/arrival airports for the 0th leg, which is an incorrect way of
doing it. Now make sure to gather airports using a leg number and also a boolean to
denote departure or arrival airport.

This refactor cleaned up some ugly code duplication, too.