Refactor flight search/filter meta-data retrieval
This moves all of generation/retrieval of the data necessary
to display pretty tablet flight filters in to FlightTripQuery.

The main benefit is that it cuts down on unnecessary cycles
in processing the FlightTrips, but also this means the filter
UI code is cleaner because it doesn't have to perform the
calculations.

This largely cleans up the concerns from Phab:
https://phabricator.mobiata.com/rEBAD66e26a8a5c6bc67355da9f62eccb41c4b26d8eef