Calculate airlines present after filtering by stops/airports
This data is used in the airline filter to determine whether or
not the airline is "enabled", i.e interactable or greyed out. If
an airline does not show up after filtering by stops and airports,
we will grey out the airline but still show its cheapest flight had
it not been filtered out.