1885 - flight leg selection, mark as viewed
When selecting the first leg of a flight via "Add for $120" button,
show the selected leg with a checkmark at the top the of the list for
the return leg (if there's a return leg, i.e roundtrip flight). This
commit does the following:

1. Create a new FlightLegSummarySection XML that contains a checkmark
instead of a price.
2. When selecting a flight, make sure to bind it to the animation row
in the FlightDetails fragment.
3. Also make sure to bind it to the section that is present in the
entire results activity (which comes in to play immediately after
the animation). This binding takes place in a state finalizer method
to ensure this gets binded on the event click, and also after rotation
(or any other means that could send us to this state).