Make the hotels list a state provider, and make it stop proxying scroll events to the activity.
This is a breaking change, as scrolling the list will never allow us to enter hotels mode (or do anything at all).
Basically this is a stepping stone commit, we dont want to proxy anything to the activity, because we want things
to be able to work independantly of what is above them, the hope here is that the HotelsController will react to
the state changes of the hotels list, and the activity can become a hotels state listener. This will be a handful of
commits down the road, but it keeps things nice and consistant, and it makes things less interdependant. E.g. In the
future I wont tell the activity anything, but rather I will just broadcast my state, and other things can react to it
however they want.